feat: enhance build process with folder support and improve directory handling

+++ _build_scripts/build.ts
@@ -2,7 +2,12 @@ import { buildSrcFile } from "./build-src.ts";
 import { buildComponent } from "./build-component.ts";
 
 async function cleaerDistFolder() {
-    await Deno.remove("dist", { recursive: true });
+    const hasDistFolder = await Deno.stat("dist").catch(() => null);
+
+    if (hasDistFolder) {
+        await Deno.remove("dist", { recursive: true });
+    }
+
     await Deno.mkdir("dist");
     await Deno.mkdir("dist/src");
 }
@@ -10,23 +15,27 @@ async function cleaerDistFolder() {
 await cleaerDistFolder();
 
 // Components
-// await buildComponent("activity-state");
-// await buildComponent("app-header");
-// await buildComponent("dynamic-rows");
-// await buildComponent("dynamic-columns");
-await buildComponent("material-icon", ["icons"]);
-// await buildComponent("menu");
+await buildComponent("activity-state");
+await buildComponent("app-header");
+await buildComponent("dynamic-rows");
+await buildComponent("dynamic-columns");
+await buildComponent("material-icon", "", ["icons"]);
+await buildComponent("divider-item", "menu");
+await buildComponent("menu-container", "menu");
+await buildComponent("menu-group", "menu");
+await buildComponent("menu-item", "menu");
+await buildComponent("menu-label", "menu");
+await buildComponent("toast-notification");
+await buildComponent("tool-bar");
+await buildComponent("tree-view");
 // await buildComponent("ollama-ui");
-// await buildComponent("toast-notification");
-// await buildComponent("toolbar");
-// await buildComponent("tree-view");
 
 // System Files
-// await buildSrcFile("system/assert.js");
-// await buildSrcFile("system/events-manager.js");
-// await buildSrcFile("system/logger.js");
+await buildSrcFile("system/assert.js");
+await buildSrcFile("system/events-manager.js");
+await buildSrcFile("system/logger.js");
 
 // Validation Files
-// await buildSrcFile("validate/conditions.js");
+await buildSrcFile("validate/conditions.js");
 
-  
+Deno.kill(Deno.pid); // Exit the process
